parser/pyproject.toml
[tool.poetry]
name = "parser_all"
version = "0.1.0"
description = ""
authors = ["DmitryMogilnikov <d.mogqs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.2.2"
docx = "^0.2.4"
"pdfminer.six" = "^20191110"
python-docx = "^0.8.11"
beautifulsoup4 = "^4.8.0"
textract = "^1.6.5"
python-djvulibre = "^0.8.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
